+++ dandelion/__init__.py
@@ -1,6 +1,6 @@
 import argparse
 
-__version__ = '0.7.0'
+__version__ = '0.7.1'
 
 def print_separator(text, width=70):
     border = "╔" + "═" * (width-2) + "╗"